Compile + Improvements (On Linux & Mac)

+++ PocketBukkit/start.sh
@@ -28,10 +28,14 @@ if [ "$PB" == "" ]; then
   if [ -f ./PocketBukkit.jar ]; then
     PB="jar"
   else
-    if [ -f ./src/net/pocketbukkit/PocketBukkit.java ]; then
-      PB="compile"
+    if [ -f ./src/net/pocketbukkit/PocketBukkit.class ]; then
+      PB="runsrc"
     else
-      PB="download"
+      if [ -f ./src/net/pocketbukkit/PocketBukkit.java ]; then
+        PB="compile"
+      else
+        PB="download"
+      fi
     fi
   fi
 fi
@@ -40,14 +44,26 @@ if [ "$PB" == "jar" ]; then
   java -jar PocketBukkit.jar
 else
   if [ "$PB" == "compile" ]; then
-    #TODO: Compile PocketBukkit
+    javac src/net/pocketbukkit/*.java
+    javac src/net/pocketbukkit/*/*.java
+    javac src/net/pocketbukkit/*/*/*.java
+    java src/net/pocketbukkit/PocketBukkit.class
   else
     if [ "$PB" == "download" ]; then
       echo "[INFO] [PocketBukkit] PocketBukkit downloader for Linux & Mac"
       echo "[1/3] Cleaning directory..."
-      rm -r start.bat
-      rm -r PocketBukkit.jar
-      rm -r -f src/
+      if [ -f ./start.bat ]; then
+        rm -r start.bat
+      fi
+      if [ -f ./PocketBukkit.jar ]; then
+        rm -r PocketBukkit.jar
+      fi
+      if [ -f ./src/net/pocketbukkit/PocketBukkit.java ]; then
+        rm -r -f src/
+      fi
+      if [ -f ./src/net/pocketbukkit/PocketBukkit.class ]; then
+        rm -r -f src/
+      fi
       echo "Done!"
       echo "[2/3] Downloading latest PocketBukkit JAR..."
       download "http://pocketbukkit.net/download/latest"
@@ -55,8 +71,12 @@ else
       echo "[3/3] Starting PocketBukkit..."
       java -jar PocketBukkit.jar
     else
-      echo "[FATAL] [PocketBukkit] Unable to determine required operation!"
-      exit 1
+      if [ "$PB" == "runsrc" ]; then
+        java src/net/pocketbukkit/PocketBukkit.class
+      else
+        echo "[FATAL] [PocketBukkit] Unable to determine required operation!"
+        exit 1
+      fi
     fi
   fi
 fi
